Move sirsi test into seperate file.
